- Rename base nn model
